Move RecentAppsThumbnailHidingLifecycleTracker into lib instead
