+++ lib/src/main/java/co/nimblehq/recentapps/thumbnailhiding/HardwareKeyWatcher.java
@@ -56,7 +56,7 @@ public class HardwareKeyWatcher {
             if (Intent.ACTION_CLOSE_SYSTEM_DIALOGS.equals(action)) {
                 String reason = intent.getStringExtra(SYSTEM_DIALOG_REASON_KEY);
                 if (reason != null) {
-                    Log.e(TAG, "action:" + action + ",reason:" + reason);
+                    Log.i(TAG, "action:" + action + ", reason:" + reason);
                     if (mListener != null) {
                         switch (reason) {
                             case SYSTEM_DIALOG_REASON_HOME_KEY:
